Tests: Move equal edif check into separate class, add large test marker annotation (#425)
* move reusable part of TestBinaryEDIF into separate class


* add marker annotation for large tests


* move md5 from string into separate function
