+++ test/src/com/xilinx/rapidwright/rwroute/TestRWRoute.java
@@ -22,11 +22,13 @@
  
 package com.xilinx.rapidwright.rwroute;
 
+import com.xilinx.rapidwright.support.LargeTest;
 import com.xilinx.rapidwright.support.RapidWrightDCP;
 import org.junit.jupiter.api.Test;
 
 import com.xilinx.rapidwright.design.Design;
 
+@LargeTest
 public class TestRWRoute {
 	/**
 	 * Tests the non-timing driven full routing, i.e., RWRoute running in its wirelength-driven mode.
